+++ sdk/appcontainers/arm-appcontainers/src/api/namespaces/index.ts
@@ -1,0 +1,5 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export { checkNameAvailability } from "./operations.js";
+export type { NamespacesCheckNameAvailabilityOptionalParams } from "./options.js";