+++ sdk/storage/arm-storage/samples/v20/javascript/fileServicesGetServiceUsageSample.js
@@ -1,0 +1,25 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+const { StorageManagementClient } = require("@azure/arm-storage");
+const { DefaultAzureCredential } = require("@azure/identity");
+
+/**
+ * This sample demonstrates how to gets the usage of file service in storage account including account limits, file share limits and constants used in recommendations and bursting formula.
+ *
+ * @summary gets the usage of file service in storage account including account limits, file share limits and constants used in recommendations and bursting formula.
+ * x-ms-original-file: 2025-08-01/FileServicesGetUsage.json
+ */
+async function getFileServiceUsage() {
+  const credential = new DefaultAzureCredential();
+  const subscriptionId = "00000000-1111-2222-3333-444444444444";
+  const client = new StorageManagementClient(credential, subscriptionId);
+  const result = await client.fileServices.getServiceUsage("res4410", "sto8607");
+  console.log(result);
+}
+
+async function main() {
+  await getFileServiceUsage();
+}
+
+main().catch(console.error);