+++ sdk/storage/arm-storage/src/api/localUsers/index.ts
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export { regeneratePassword, listKeys, list, $delete, createOrUpdate, get } from "./operations.js";
+export type {
+  LocalUsersRegeneratePasswordOptionalParams,
+  LocalUsersListKeysOptionalParams,
+  LocalUsersListOptionalParams,
+  LocalUsersDeleteOptionalParams,
+  LocalUsersCreateOrUpdateOptionalParams,
+  LocalUsersGetOptionalParams,
+} from "./options.js";