+++ sdk/keyvault/keyvault-keys/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- Added support for service API version `2025-07-01` [#46716](https://github.com/Azure/azure-sdk-for-python/pull/46716)
+
 ### Breaking Changes
 
 ### Bugs Fixed
@@ -11,6 +13,8 @@
 - Fix dependency issue.
 
 ### Other Changes
+
+- Key Vault API version `2025-07-01` is now the default
 
 ## 4.10.0 (2025-06-10)